♻️ Add `Hyku::HomePageThemesBehavior`
This commit will move the #inject_theme_views method and the
around_action that calls it into a concern.
