+++ app/controllers/hyrax/contact_form_controller_decorator.rb
@@ -14,10 +14,9 @@ module Hyrax
     # Adds Hydra behaviors into the application controller
     include Blacklight::SearchContext
     include Blacklight::AccessControls::Catalog
+    include Hyku::HomePageThemesBehavior
 
     prepended do
-      # OVERRIDE: Adding inject theme views method for theming
-      around_action :inject_theme_views
       before_action :setup_negative_captcha, only: %i[new create]
 
       # OVERRIDE: Add for theming
@@ -83,24 +82,6 @@ module Hyrax
       end
     rescue Blacklight::Exceptions::ECONNREFUSED, Blacklight::Exceptions::InvalidRequest
       []
-    end
-
-    # OVERRIDE: Adding to prepend the theme views into the view_paths
-    def inject_theme_views
-      if home_page_theme && home_page_theme != 'default_home'
-        original_paths = view_paths
-        Hyku::Application.theme_view_path_roots.each do |root|
-          home_theme_view_path = File.join(root, 'app', 'views', "themes", home_page_theme.to_s)
-          prepend_view_path(home_theme_view_path)
-        end
-        yield
-        # rubocop:disable Lint/UselessAssignment, Layout/SpaceAroundOperators, Style/RedundantParentheses
-        # Do NOT change this line. This is calling the Rails view_paths=(paths) method and not a variable assignment.
-        view_paths=(original_paths)
-        # rubocop:enable Lint/UselessAssignment, Layout/SpaceAroundOperators, Style/RedundantParentheses
-      else
-        yield
-      end
     end
 
     def setup_negative_captcha
